feat: enable support for node v22 (#607)
* feat: enable support for node v22

* refactor: fix indentations

* feat: add node version 22 to unit-test strategy matrix

* chore: downgrade chalk and other packages so it should work for all specified node versions

---------

+++ src/utils.cpp
@@ -728,7 +728,8 @@ v8::Local<v8::Value> javaToV8(Java* java, JNIEnv* env, jobject obj, DynamicProxy
         jmethodID integer_intValue = env->GetMethodID(integerClazz, "intValue", "()I");
         jint result = env->CallIntMethod(obj, integer_intValue);
         checkJavaException(env);
-        return Nan::New<v8::Integer>(result);
+        // <-- use the Int32 factory, not the generic Integer one
+        return Nan::New<v8::Int32>(static_cast<int32_t>(result));
       }
     case TYPE_SHORT:
       {
